Add timedelta() template filter and timestamp global
Introduce a "timestamp" template global variable with the date/time at
which the "!render" command was issued, and a "timedelta()" filter that
may be used to get new timestamps calculated by adding a time duration
(in seconds, minutes, hours, etc.) to an existing one.

This allows calculating texts like "from September 20 up to October 7"
in templates.